fixing tower UI and trying to fixe base and spawn positions of enemies

+++ Assets/Scripts/Combat.cs
@@ -4,10 +4,10 @@ using UnityEngine.Networking;
 
 public class Combat : NetworkBehaviour {
 
-    public int maxHealth = 10;
+    public float maxHealth = 10;
 
     //[SyncVar]
-    protected int health;
+    protected float health;
 
     public Combat()
     {
